requirements.txt
streamlit==1.22.0
requests==2.28.2
matplotlib==3.7.5
wordcloud==1.9.3
python-dotenv==1.0.0